+++ src/api/admin.ts
@@ -1,6 +1,6 @@
 import { ServiceKey } from './types/admin'
 import { Domain, ProviderService, ServiceResponse } from './types/general'
-import { getAvailableDomains, setData, getProviderKeys } from './lib/data'
+import { getAvailableDomains, setData, getProviderKeys } from '../lib/data'
 import express from 'express'
 import uuid4 from 'uuid/v4'
 import util from 'util'
