+++ src/app.ts
@@ -7,7 +7,7 @@ import { apiRouter } from './api/index'
 import https from 'https'
 import fs from 'fs'
 import tls from 'tls'
-import { getAvailableDomains } from './api/lib/data'
+import { getAvailableDomains } from './lib/data'
 import { isCorrectCredentials } from './auth'
 
 const app = express()
@@ -51,30 +51,30 @@ const listener = (): void => {
   )
 }
 
-if(process.env.NODE_ENV === 'production'){
-const server = https.createServer(
-  {
-    SNICallback: (domain, cb) => {
-      // escape characters required or readFileSync will not find file
-      const homePath = process.env.HOME
-      const secureContext = tls.createSecureContext({
-        /* eslint-disable */
-        key: fs.readFileSync(
-          `${homePath}/\.acme\.sh/*\.${domain}/*\.${domain}\.key`
-        ),
-        cert: fs.readFileSync(
-          `${homePath}/\.acme\.sh/*\.${domain}/*\.${domain}\.cer`
-        )
-        /* eslint-enable */
-      })
-      if (cb) return cb(null, secureContext)
-      return secureContext
+if (process.env.NODE_ENV === 'production') {
+  const server = https.createServer(
+    {
+      SNICallback: (domain, cb) => {
+        // escape characters required or readFileSync will not find file
+        const homePath = process.env.HOME
+        const secureContext = tls.createSecureContext({
+          /* eslint-disable */
+          key: fs.readFileSync(
+            `${homePath}/\.acme\.sh/*\.${domain}/*\.${domain}\.key`
+          ),
+          cert: fs.readFileSync(
+            `${homePath}/\.acme\.sh/*\.${domain}/*\.${domain}\.cer`
+          )
+          /* eslint-enable */
+        })
+        if (cb) return cb(null, secureContext)
+        return secureContext
+      }
+    },
+    (req, res) => {
+      res.end('hello world')
     }
-  },
-  (req, res) => {
-    res.end('hello world')
-  }
-)
+  )
 
   server.listen(443)
 }
