+++ src/providers/goDaddy.ts
@@ -1,5 +1,5 @@
 import { sendRequest } from '../helpers/httpRequest'
-import { getProviderKeys } from '../api/lib/data'
+import { getProviderKeys } from '../lib/data'
 import { Provider, ServiceResponse } from '../api/types/general'
 import { ServiceKey } from '../api/types/admin'
